# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Test API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/dashboard/monthly-summary": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get monthly summary statistics",
        "description": "Returns combined monthly invoice and customer statistics",
        "responses": {
          "200": {
            "description": "Monthly summary data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/MonthlySummaryItem"
                  }
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store, max-age=0"
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/v1/dashboard/weekly-summary": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get weekly summary statistics",
        "description": "Returns combined weekly invoice and customer statistics",
        "responses": {
          "200": {
            "description": "Weekly summary data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/WeeklySummaryItem"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/api/v1/invoices/generate-number": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "Generate a unique invoice number",
        "description": "Generates a unique invoice number with retry logic",
        "responses": {
          "200": {
            "description": "Generated invoice number",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "invoiceNumber": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "no-store, max-age=0"
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/v1/invoices/{id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "Get invoice by ID",
        "description": "Retrieves a specific invoice with details and user information",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Invoices"
        ],
        "summary": "Update invoice by ID",
        "description": "Updates an existing invoice and recalculates totals",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Invoices"
        ],
        "summary": "Delete invoice by ID",
        "description": "Deletes a specific invoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/v1/invoices/{id}/toggle-accounting": {
      "patch": {
        "tags": [
          "Invoices"
        ],
        "summary": "Toggle accounting flag",
        "description": "Toggles the accounting flag for a specific invoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounting flag toggled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "vatNumber": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "MonthlySummaryItem": {
        "type": "object",
        "properties": {
          "totalAmount": {
            "type": "number",
            "format": "float"
          },
          "totalTtcAmount": {
            "type": "number",
            "format": "float"
          },
          "totalCustomers": {
            "type": "integer"
          }
        },
        "required": [
          "totalAmount",
          "totalTtcAmount",
          "totalCustomers"
        ]
      },
      "WeeklySummaryItem": {
        "type": "object",
        "properties": {
          "totalAmount": {
            "type": "number",
            "format": "float"
          },
          "totalTtcAmount": {
            "type": "number",
            "format": "float"
          },
          "totalCustomers": {
            "type": "integer"
          }
        },
        "required": [
          "totalAmount",
          "totalTtcAmount",
          "totalCustomers"
        ]
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "float"
          },
          "totalVatAmount": {
            "type": "number",
            "format": "float"
          },
          "totalTtcAmount": {
            "type": "number",
            "format": "float"
          },
          "flag_accounting": {
            "type": "boolean"
          },
          "User": {
            "$ref": "#/components/schemas/User"
          },
          "invoiceDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDetail"
            }
          }
        }
      },
      "InvoiceDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "lineNumber": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          },
          "discount": {
            "type": "number",
            "format": "float"
          },
          "unitPrice": {
            "type": "number",
            "format": "float"
          },
          "vatBaseAmount": {
            "type": "number",
            "format": "float"
          },
          "vatAmount": {
            "type": "number",
            "format": "float"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          },
          "item": {
            "$ref": "#/components/schemas/Item"
          }
        }
      },
...
      "InvoiceItemCreate": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "discount": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "itemId",
          "quantity"
        ]
      },
      "InvoiceUpdate": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemCreate"
            }
          }
        },
        "required": [
          "userId",
          "items"
        ]
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "itemNumber": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "retailPrice": {
            "type": "number",
            "format": "float"
          },
          "purchasePrice": {
            "type": "number",
            "format": "float"
          },
          "stockQuantity": {
            "type": "integer"
          },
          "minQuantity": {
            "type": "integer"
          },
          "vat": {
            "$ref": "#/components/schemas/Vat"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "itemClass": {
            "$ref": "#/components/schemas/ItemClass"
          },
          "itemTaxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemTax"
            }
          }
        }
      },
...
      "Vat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "vatPercent": {
            "type": "number",
            "format": "float"
          },
          "vatType": {
            "type": "string"
          },
          "country": {
            "$ref": "#/components/schemas/Country"
          }
        }
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "countryCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ItemTax": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "itemId": {
            "type": "integer"
          },
          "taxId": {
            "type": "integer"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "utax": {
            "$ref": "#/components/schemas/Utax"
          }
        }
      },
      "Utax": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "rate": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "ItemClass": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
...
      "Unit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}